Move queue scheme to schemes folder
